Split typed type core
